refactor gui and move creation of LoggedInUserSingleton to interactor
