+++ src/main/java/intefaceAdapters/userRegister/UserCreationFailed.java
@@ -1,4 +1,4 @@
-package screens.userRegisterScreen;
+package intefaceAdapters.userRegister;
 
 public class UserCreationFailed extends RuntimeException {
     public UserCreationFailed(String error) {
