+++ src/main/java/intefaceAdapters/userRegister/UserRegisterController.java
@@ -1,4 +1,4 @@
-package screens.userRegisterScreen;
+package intefaceAdapters.userRegister;
 
 
 import useCases.userRegister.UserRegisterInputBoundary;
